tests/wpt/web-platform-tests/payment-handler/can-make-payment-event-constructor.https.html

<!doctype html>
<meta charset="utf-8">
<title>Test for CanMakePaymentEvent Constructor</title>
<link rel="help" href="https://w3c.github.io/payment-handler/#dom-canmakepaymentevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

test(() => {
  assert_false('CanMakePaymentEvent' in window);
}, 'CanMakePaymentEvent constructor must not be exposed in window');

</script>
